+++ Dijkstra.h
@@ -8,7 +8,7 @@ class Dijkstra
     Cell* start;
     Cell* end;
     priority_queue<Cell*, vector<Cell*>, CellComparator> frontier; // Min-heap
-    float oldCost[NumVerticalCells][NumHorizontalCells];
+    double oldCost[NumVerticalCells][NumHorizontalCells];
     bool done;
     vector<Cell*> path;
