+++ GreedyBFS.h
@@ -10,7 +10,7 @@ class GreedyBFS
     Cell* start;
     Cell* end;
     priority_queue<Cell*, vector<Cell*>, CellComparator> frontier;
-    float oldCost[NumVerticalCells][NumHorizontalCells];
+    double oldCost[NumVerticalCells][NumHorizontalCells];
     bool done;
     vector<Cell*> path;
